Processes with stable diffusion

+++ Pokora/Model/Effect.swift
@@ -13,9 +13,9 @@ struct Effect: Identifiable {
     var id = UUID()
     var startFrame: Int
     var endFrame: Int
-    
+
     // for now all effects are Stable Diffusion, but I could refactor this
-    var strength: Double = 0.0
+    var strength: Float = 0.0
     var seed: UInt32 = globalSeed
     var prompt: String = "a cyberpunk cityscape"
 }
